+ src_query

+++ operators/oracle_to_postgres.py
@@ -31,38 +31,50 @@ class OracleToPostgresOperator(BaseOperator):
             with closing(OracleHook(oracle_conn_id=self.oracle_conn_id).get_conn()) as src_conn:
                 with closing(PostgresHook(postgres_conn_id=self.postgres_conn_id).get_conn()) as tgt_conn:
 
-                    select_query = "select {fields} from {src_schema}.{src_table}".format(**self.params)
-                    if self.params['use_conditions']:
-                        select_query += " where {conditions}".format(**self.params)
-                    logging.info(f"Run query: \"{select_query}\"")
+                    if self.params['is_active']: 
 
-                    # открытие курсоров на чтение и запись
-                    src_cursor = src_conn.cursor("serverCursor")
-                    tgt_cursor = tgt_conn.cursor()
-                    src_cursor.execute(select_query)
-
-                    # очистка Stage-таблицы
-                    tgt_cursor.execute("truncate table {tgt_schema}.{target_table_prefix}{tgt_table}".format(**self.params))
-
-                    # обработка результата запроса
-                    batch_count = 0
-                    while True:
-                        logging.info(f"Processing batch:\t{batch_count},\tsize: {self.batch_size}")
-                        records = src_cursor.fetchmany(self.batch_size)
-                        
-                        if records:
-                            execute_values(  # вставка батча данных
-                                tgt_cursor,
-                                "INSERT INTO {tgt_schema}.stg_{tgt_table} ({fields}) VALUES %s".format(**self.params),
-                                records
-                            )
+                        # если src_query не пустое, используем его, иначе конструируем запрос из полей fields, src_schema, src_table и conditions
+                        if self.params['src_query']: 
+                            select_query = self.params['src_query']
                         else:
-                            logging.info("Передача данных из таблицы {src_schema}.{src_table} завершена".format(**self.params))
-                            break
+                            select_query = "select {fields} from {src_schema}.{src_table}".format(**self.params)
+                            
+                            if self.params['use_conditions']:
+                                select_query += " where {conditions}".format(**self.params)
+                            logging.info(f"Run query: \"{select_query}\"")
 
-                        batch_count += 1
-                    
-                    tgt_conn.commit()
+                        # открытие курсоров на чтение и запись
+                        src_cursor = src_conn.cursor("serverCursor")
+                        tgt_cursor = tgt_conn.cursor()
 
-                    src_cursor.close()
-                    tgt_cursor.close()
+                        # выполнение запроса
+                        src_cursor.execute(select_query)
+
+                        # очистка Stage-таблицы
+                        tgt_cursor.execute("truncate table {tgt_schema}.{target_table_prefix}{tgt_table}".format(**self.params))
+
+                        # обработка результата запроса
+                        batch_count = 0
+                        while True:
+                            logging.info(f"Processing batch:\t{batch_count},\tsize: {self.batch_size}")
+                            records = src_cursor.fetchmany(self.batch_size)
+                            
+                            if records:
+                                execute_values(  # вставка батча данных
+                                    tgt_cursor,
+                                    "INSERT INTO {tgt_schema}.{target_table_prefix}{tgt_table} ({fields}) VALUES %s".format(**self.params),
+                                    records
+                                )
+                            else:
+                                logging.info("Передача данных из таблицы {src_schema}.{src_table} завершена".format(**self.params))
+                                break
+
+                            batch_count += 1
+                        
+                        tgt_conn.commit()
+
+                        src_cursor.close()
+                        tgt_cursor.close()
+
+                    else:
+                        logging.info("Таблица {src_schema}.{src_table} не активна в метаданных, пропускаем...".format(**self.params))
